combine FederationJoin and ActivitySender
they both allow a server to act as a client for another server in the federation
