+++ src/main/java/de/tuberlin/tkn/lit/LitApplication.java
@@ -1,20 +1,18 @@
 package de.tuberlin.tkn.lit;
 
-import de.tuberlin.tkn.lit.FederationJoin;
+import de.tuberlin.tkn.lit.processing.IFederationClient;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
-
 import org.springframework.boot.context.event.ApplicationReadyEvent;
 import org.springframework.context.event.*;
-
 import org.springframework.beans.factory.annotation.Autowired;
 
 @SpringBootApplication
 public class LitApplication {
 	
 	@Autowired
-	FederationJoin federationJoin;
+	IFederationClient federationClient;
 
 	public static void main(String[] args) {
 		SpringApplication.run(LitApplication.class, args);
@@ -23,7 +21,7 @@ public class LitApplication {
 	@EventListener(ApplicationReadyEvent.class)
 	public void joinFederation() {
 		
-		String[] args = { "http://localhost:8080" };
+		String[] args = { "http://localhost:8080" }; // TODO : this should be from the actual arguments
 
 		// get command line arguments
 		String knownMember = args.length == 1 ? args[0] : "";
@@ -32,7 +30,7 @@ public class LitApplication {
 		// supplies the adress of a member
 		if (!knownMember.isEmpty()) {
 			System.out.print("Try joining a federation \n");
-			federationJoin.handleJoinFederation(knownMember);
+			federationClient.handleJoinFederation(knownMember);
 		}
 	}
 }
