+++ src/main/java/de/tuberlin/tkn/lit/processing/IFederationClient.java
@@ -6,7 +6,8 @@ import de.tuberlin.tkn.lit.model.activitypub.activities.Activity;
 import de.tuberlin.tkn.lit.model.activitypub.core.LinkOrObject;
 import de.tuberlin.tkn.lit.model.activitypub.core.OrderedCollection;
 
-public interface IActivitySender{
+public interface IFederationClient{
 	Future<Boolean> send(Activity activity, LinkOrObject sendTo);
 	Future<OrderedCollection> getOutbox(LinkOrObject getFrom);
+	void handleJoinFederation(String knownMember);
 }
